+++ tests/testthat/testCavg.R
@@ -3,8 +3,8 @@ library(dplyr)
 library(campsis)
 context("Test Cavg metric")
 
-testFolder <- ""
-source(paste0(testFolder, "testUtils.R"))
+testFolder <-  file.path(getwd(), test_path())
+source(file.path(testFolder, "testUtils.R"))
 
 test_that("Dataset 1 - cavg at steady state", {
   ds <- dataset1()